feat: server-side evidence extraction with frontend fallback (Phases 1-3)
Backend now sends extracted evidence in SSE events, eliminating
frontend's need to duplicate 12 evidence extraction files. Frontend
uses server data when available, falls back to client-side extraction.

- Add shared evidence types (Decision, Citation, VaultDocument)
- Export extractFromToolResult/extractNormsFromAnswer as public API
- Add evidence field to tool_result SSE events
- Add norms field to answer SSE events
- Add cumulative evidence_update SSE event type
- Frontend MCPService: parse new event types
- Frontend useAIChatStream: prefer server evidence with client fallback
