+++ packages/shared/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './cost';
+export * from './evidence';
 export * from './http';
